+++ src/main/java/calculator/MultiplyOperator.java
@@ -2,7 +2,7 @@ package calculator;
 
 public class MultiplyOperator implements Operator {
     @Override
-    public int operate(int num1, int num2) {
+    public double operate(double num1, double num2) {
         return num1 * num2;
     }
 }
