+++ src/main/java/calculator/OperatorType.java
@@ -22,7 +22,7 @@ public enum OperatorType {
         return Arrays.stream(OperatorType.values()).filter(t -> t.operatorChar == operator).findFirst();
     }
 
-    public int operate(int num1, int num2) {
+    public double operate(double num1, double num2) {
         return operator.operate(num1, num2);
     }
 }
